elasticapm: split span bits into span.go
